+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Physics backend plugins are now in charge of preventing `TnuaSystemSet` from
+  running while the physiscs backend is paused. Users no longer need to do it.
+
 ## 0.6.1 - 2023-06-04
 ### Fixed
 - Fix jump shortening not working with takeoff gravity